This commit was manufactured by cvs2svn to create tag 'SpecTcl-2-1_022'.

+++ SpecTcl/Display/Xamine.h
@@ -348,8 +348,8 @@ typedef union {
 
 
 typedef struct {
-                 unsigned short xchans;
-		 unsigned short ychans;
+                 unsigned int xchans;
+		 unsigned int ychans;
 	       } spec_dimension;	/* Describes the channels in a spectrum. */
 
 typedef char spec_title[72];
